update brushes to plugin scripting brushes
Andy ran the "Export brushes for toolkit" editor tool in the Open Brush Unity project, while on the feature/plugin-scripting branch, and gave the file to me. I copied the new shaders and materials to this project.

We did this because the Sketch Playback feature for the Unity SDK requires the ClipStart,ClipEnd parameters for the brush shaders.